requirements.txt
Flask
requests
pymilvus==2.4.1
sentence-transformers
python-dotenv
ibm-watson
ibm-watson-machine-learning
flassger